requirements.txt
flask
Pillow
torch>=2.2.0
transformers>=4.37.0
accelerate>=0.27.0
sentencepiece
python-dotenv
safetensors
datasets
--find-links https://download.pytorch.org/whl/torch_stable.html
